agent-server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agent-server"
version = "0.1.0"
description = "Localhost AI Agent Backend — FastAPI + Claude API + MCP"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    # Claude API
    "anthropic>=0.28.0",
    # MCP
    "mcp>=1.0.0",
    # DB (PostgreSQL)
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Crypto (Vault)
    "cryptography>=42.0.0",
    # HTTP
    "httpx>=0.27.0",
    # Config / Env
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    # Email
    "aiosmtplib>=3.0.0",
    # Token counting
    "tiktoken>=0.7.0",
    # SSE
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
